Module move_vm_state::execution_context

Structs

SystemExecutionContext

TransactionExecutionContext
A TransactionExecutionContext holds the mutable data that needs to be persisted from one section of the transaction flow to another. Because of this, this is the only data that can both be mutated, and persist between interpretation instances.

Traits

ExecutionContext
An ExecutionContext represents mutable state that is retained in-memory between invocations of the Move VM.