Struct move_vm_state::execution_context::TransactionExecutionContext

pub struct TransactionExecutionContext<'txn> { /* fields omitted */ }

A TransactionExecutionContext holds the mutable data that needs to be persisted from one section of the transaction flow to another. Because of this, this is the only data that can both be mutated, and persist between interpretation instances.

Methods

impl<'txn> TransactionExecutionContext<'txn>

The transaction

pub fn new(
gas_left: GasUnits<GasCarrier>,
data_cache: &'txn dyn RemoteCache
) -> Self